% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxCox.R
\name{boxCox}
\alias{boxCox}
\title{Box-Cox Expression}
\usage{
boxCox(x, lambda)
}
\arguments{
\item{x}{numeric vector of data}

\item{lambda}{numeric value}
}
\value{
numeric value; l(lambda) for the given lambda
}
\description{
Box-Cox Expression
}
\examples{
boxCox(rnorm(50), 2)
}
