% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxCoxMax.R
\name{boxCoxMax}
\alias{boxCoxMax}
\title{Box-Cox Expression Solution}
\usage{
boxCoxMax(x, lambdaMin = -3, lambdaMax = 5, increment = 0.001)
}
\arguments{
\item{x}{numeric vector of data}

\item{lambdaMin}{numeric value; minimum value of lambda}

\item{lambdaMax}{numeric value; maximum value of lambda}

\item{increment}{numeric value; delta}
}
\value{
named list; solution (lambda value that maximizes the Box-Cox expression), maximum (max value of expression), and a plot (l(lambda) vs lambda)
}
\description{
Box-Cox Expression Solution
}
\examples{
boxCoxMax(rnorm(50))
}
