% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirstShiny.R
\name{myfirstShiny}
\alias{myfirstShiny}
\title{My First Shiny App}
\usage{
myfirstShiny()
}
\value{
interactive html shiny browser
}
\description{
My First Shiny App
}
\examples{
\dontrun{MATH5793YUT::myfirstShiny()}
}
