% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPCA.R
\name{myPCA}
\alias{myPCA}
\title{Principal Component Analysis}
\usage{
myPCA(Sigma)
}
\arguments{
\item{Sigma}{variance-covariance matrix}
}
\value{
named list with PCA coefficients, correlation coeficients, and population variance proportion for non-standardized and standardized data
}
\description{
Principal Component Analysis
}
\examples{
myPCA(matrix(c(1,-2,0, -2,5,0, 0,0,2), ncol = 3, byrow =TRUE))
}
