% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptn4.29.R
\name{ptn4.29}
\alias{ptn4.29}
\title{Proportion Test of Normality (JW 4-29)}
\usage{
ptn4.29(data)
}
\arguments{
\item{data}{data.frame with numeric variables}
}
\value{
named list containing the proportion of the data within 1 and 2 standard deviations as well as the test result for each variable Xi in data
}
\description{
Proportion Test of Normality (JW 4-29)
}
\examples{
ptn4.29(data=data.frame(x=rnorm(50), y=rchisq(50,1)))
}
