% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQ.R
\name{rQ}
\alias{rQ}
\title{Q-Q Plot Correlation Coefficient}
\usage{
rQ(Xi)
}
\arguments{
\item{Xi}{numeric vector; continuous univariate data}
}
\value{
numeric value; the correlation coefficient (equation 4-31 from JW)
}
\description{
Q-Q Plot Correlation Coefficient
}
\examples{
rQ(runif(50)); rQ(rnorm(30))
}
