% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMVN.R
\name{shinyMVN}
\alias{shinyMVN}
\title{Shiny App to Test Multivariate Normality}
\usage{
shinyMVN()
}
\value{
interactive html shiny browser
}
\description{
Shiny App to Test Multivariate Normality
}
\examples{
\dontrun{MATH5793YUT::shinyMVN()}
}
