% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T4.3}
\alias{T4.3}
\title{Four Measurements of Stiffness}
\format{
A data frame with 30 rows and 5 variables:
\describe{
  \item{x1}{measure 1 of stiffness}
  \item{x2}{measure 2 of stiffness}
  \item{x3}{measure 3 of stiffness}
  \item{x4}{measure 4 of stiffness}
  \item{dsq}{squared distances}

}
}
\source{
{Applied Multivariate Statistical Analysis by Johnson & Wichern}
}
\usage{
T4.3
}
\description{
A dataset containing four measurements of stiffness for 30 boards, as well as squared distances.
}
\keyword{datasets}
