% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tsq.test.R
\name{Tsq.test}
\alias{Tsq.test}
\title{T-Squared Test and Plot}
\usage{
Tsq.test(data, mu0, alpha = 0.05)
}
\arguments{
\item{data}{a data.frame with p variables and n observations}

\item{mu0}{a numeric vector, length p}

\item{alpha}{a numeric value between 0 and 1}
}
\value{
named list
}
\description{
T-Squared Test and Plot
}
\examples{
Tsq.test(data=data.frame(rnorm(50), rnorm(50)), mu0=c(-0.5, 1.2))
}
